move the baseparser.py inside the base submodule
